ListItem download function complete listitem background bug:Could not find a part of the path "D:\unitydemo\Anywhere\AnywhereUP\Assets\StreamingAssets\ResCache\Download\Texture\TestPreview.png".

+++ AnywhereUP/Assets/Scripts/OSS/GetObject.cs
@@ -130,7 +130,7 @@ namespace Aliyun.OSS
                             fs.Write(buf, 0, length);
                             downloadLen += length;
                             Debug.LogError("download length:" + downloadLen);
-                            m_DownLoadProgress = downloadLen / result.Metadata.ContentLength;
+                            m_DownLoadProgress = (float)downloadLen / result.Metadata.ContentLength;
                             m_DownLoadState = DownLoadState.DOWNLOADING;
                         } while (length != 0);
                     }
